Major overhaul of the 'ApiUsage' class, begin separation into an examples submodule; improve and fix 'CollectionUtils' and 'BtcdClient'

+++ src/main/java/com/neemre/btcdcli4j/client/BtcdClient.java
@@ -22,7 +22,7 @@ public interface BtcdClient {
 	
 	Boolean getGenerate();
 	
-	Integer getHashesPerSec();
+	Long getHashesPerSec();
 	
 	Info getInfo();
